chore(agoric-cli): rename `lib` to `src`
