+++ packages/agoric-cli/bin/agoric
@@ -1,1 +1,1 @@
-../lib/entrypoint.js
+../src/entrypoint.js
