+++ packages/agoric-cli/package.json
@@ -3,7 +3,7 @@
   "version": "0.13.19",
   "description": "Manage the Agoric Javascript smart contract platform",
   "type": "module",
-  "main": "lib/main.js",
+  "main": "src/main.js",
   "bin": "bin/agoric",
   "files": [
     "lib",
