+++ packages/agoric-cli/test/test-main.js
@@ -8,7 +8,7 @@ import test from 'ava';
 import fs from 'fs';
 import anylogger from 'anylogger';
 
-import main from '../lib/main.js';
+import main from '../src/main.js';
 
 test('sanity', async t => {
   const stubAnylogger = () => {
